pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mantisanalysis"
version = "0.1.0"
description = "MantisCam RGB-NIR analysis suite — USAF resolution, FPN, Depth-of-Field"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Zhongmin Zhu", email = "j@polarxphotonics.com" }]
maintainers = [{ name = "BioSensors Lab @ UIUC" }]
keywords = ["imaging", "USAF", "MTF", "FPN", "depth-of-field",
            "RGB-NIR", "MantisCam", "GSense"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "h5py>=3.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "Pillow>=10.0",
    "tifffile>=2023.0",
    "PySide6>=6.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-qt>=4.4",
    "ruff>=0.5",
    "mypy>=1.8",
]

[project.urls]
Homepage = "https://github.com/BioSensorsLab-Illinois/MantisAnalysis"
Repository = "https://github.com/BioSensorsLab-Illinois/MantisAnalysis"
Issues = "https://github.com/BioSensorsLab-Illinois/MantisAnalysis/issues"

[project.scripts]
mantisanalysis = "mantisanalysis.app:main_argv"

[project.gui-scripts]
"MantisAnalysis-gui" = "mantisanalysis.app:main_argv"

[tool.setuptools]
packages = ["mantisanalysis", "mantisanalysis.modes"]
include-package-data = true

[tool.setuptools.package-data]
mantisanalysis = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["outputs", "build", "dist", ".venv"]

[tool.ruff.lint]
# Conservative: no auto-formatting refactors imposed on the existing code.
# Focus on correctness lints + import hygiene.
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
    "E501",   # line too long — many existing lines are wider, address later
    "E731",   # lambda assignments — used in Qt callback wiring
    "B008",   # function call as default arg — Qt patterns
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]  # asserts are fine in tests

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "gui: marks tests that require a Qt display (deselect with '-m \"not gui\"')",
    "headless: matplotlib Agg-backend figure tests; safe in CI",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
no_implicit_optional = true
# Run on the package only at first; widen later.
files = ["mantisanalysis"]
